Lookup reward pool account from db instead of meteora (#13441)
### Description
Instead of using the `leftoverReceiver` from onchain meteora config,
let's use the values we indexed from the reward manager init
instruction, because for older configs the leftoverReceiver was set to
the creator wallet instead of the reward pool.

### How Has This Been Tested?

Ran local solana relay, hit endpoint manually for `PHUTURES` and
confirmed the reward pool address was correct.

```
{"level":"INFO","time":"2025-11-17T21:49:19.796Z","name":"solana-relay","message":"Added transfer instructions","userTransferAmount":"5479451054794520","rewardsPoolTransferAmount":"5479451054794520","rewardsPoolAddress":"3PD2E2mvjHp6MJuj3MNTY3rsYoWQ1UJ6UT9T5d9DGkQh","from":"BDYimsg6pUjunqsKL37jhEuyzFW3FVRBjxvDUeqzTYgQ","to":"GrWNH9qfwrvoCEoTm65hmnSh4z3CD96SfhtfQY6ZKUfY"}
```